makefile: hotfix for kaniko on cgroup v2 hosts
This is a temporary workaround that enables kaniko to run on cgroup v2
enabled hosts. Due to an upstream issue, kaniko fails to detect that it
is indeed being run from a container.
Solution introduced here is to force kaniko to run regardless.

as seen in: https://git.dotya.ml/wanderer/docker-fedora-hugo/commit/d728f4cd6066e314687174db3730990fc9030ddd

[skip ci]